lugx-microservices/analytics-service/requirements.txt
flask
clickhouse-connect
flask_cors
python-dotenv
prometheus_client
